move unpackTimeSeries to datalib.py
