+++ moira/graphite/datalib.py
@@ -98,10 +98,31 @@ class TimeSeries(list):
             'values': list(self),
         }
 
-# Data retrieval API
-def extract(value):
-    parts = value.split()
-    return float(parts[1])
+
+def unpackTimeSeries(dataList, retention, startTime, endTime, allowRealTimeAlerting):
+
+    def getTimeSlot(timestamp):
+        return int((timestamp - startTime) / retention)
+
+    valuesList = []
+    for data in dataList:
+        points = {}
+        for value, timestamp in data:
+            points[getTimeSlot(timestamp)] = float(value.split()[1])
+
+        lastTimeSlot = getTimeSlot(endTime)
+
+        values = []
+        # note that right boundary is exclusive
+        for timeSlot in range(0, lastTimeSlot):
+            values.append(points.get(timeSlot))
+
+        lastPoint = points.get(lastTimeSlot)
+        if allowRealTimeAlerting and lastPoint is not None:
+            values.append(lastPoint)
+
+        valuesList.append(values)
+    return valuesList
 
 
 @defer.inlineCallbacks
